==> Codebase/2_preprocess_data.ipynb <==
import json
import pandas as pd

version = 11

with open("codebook.json", "r") as codebook_file:
    codebook = json.load(codebook_file)

with open("country.json", "r") as country_file:
    countries = json.load(country_file)

df = pd.read_csv(f'ESS{version}/ESS{version}.csv')

df["id"] = df["cntry"] + '_' + df["idno"].astype(str)
df.set_index("id", inplace=True)
df = df[codebook.keys()]

df

def normalize_column(column, values, scale_type):
        
    if scale_type == 'u':
        column = column.apply(lambda x: (x - min(values)) / (max(values) - min(values)) if x not in [77, 88, 99] else x)
    elif scale_type == 'bi':
        column = column.apply(lambda x: 2 * (x - min(values)) / (max(values) - min(values)) - 1 if x not in [77, 88, 99] else x)
    
    return column

for variable in df.columns:
    values = codebook[variable]['values']
    scale_type = codebook[variable]['scale_type']
    
    
    if max(values) < 7:
        df[variable] = df[variable].replace({7: 77, 8: 88, 9: 99})

        df[variable] = normalize_column(df[variable], values, scale_type)   

df

df.to_csv("preprocessed_data.csv")
