==> Codebase/3_prepare_prompts.ipynb <==
import json
from collections import defaultdict

with open("codebook.json", "r") as codebook_file:
    codebook = json.load(codebook_file)

with open("country.json", "r") as country_file:
    countries = json.load(country_file)

def render_prompt(question, answer):
    return f"Create a statement using the given question <{question}> and its answer <{answer}>. Ensure that the response is purely a statement without bias and remains at the syntactic level. The statement should sound like it is coming directly from the respondent."

def list_of_prompts(question, pivot_response):
    prompts = []
    for answer in [pivot_response, "Refusal", "Don't know", "No answer"]:
        prompts.append(render_prompt(question, answer))

    return prompts

prompts = defaultdict(dict)
for variable in codebook.keys():

    pivot_response = codebook[variable]['categories'][-1]
    question = codebook[variable]['question']
    answers = [pivot_response, "Refusal", "Don't know", "No answer"]

    prompts[variable] = f"Create statements using the given question <{question}> and its answers <{pivot_response}>, <Refusal>, <Don't know>, and <No answer>. Ensure that the response is purely statements without bias and remains at the syntactic level. The statements should sound like it is coming directly from the respondent. Separate statements with |"
        


with open("prompts.json", "w") as prompt_file:
    json.dump(prompts, prompt_file, indent=4)
