==> Codebase/5_estimate_vectors.ipynb <==
import json
from collections import defaultdict

with open("statements.json", "r") as statements_file:
    statements = json.load(statements_file)

from sentence_transformers import SentenceTransformer

model = SentenceTransformer("BAAI/bge-m3")

with open("country.json", "r") as country_file:
    countries = json.load(country_file)

vectors = defaultdict(dict)
for variable, stats in statements.items():
    for code, country in countries.items():
        vectors[variable][code] = model.encode(stats.replace('[country]', country).split(' | '))

import pickle

with open('vectors.pkl', 'wb') as f:
    pickle.dump(vectors, f, protocol=pickle.HIGHEST_PROTOCOL)
