==> Codebase/7_projections.ipynb <==
import numpy as np
import pandas as pd
import pickle
from collections import defaultdict

with open('vectors.pkl', 'rb') as f:
    vectors = pickle.load(f)

with open('participants.pkl', 'rb') as f:
    participants = pickle.load(f)

non_normalized_df = pd.read_csv('non_normalized_data.csv')
non_normalized_df.set_index('id', inplace=True)



target_variables = ['trstprl', 'badge', 'lrscale']

distributions = defaultdict(list)

for target in target_variables:
    for participant, values in participants.items():
        country = participant[:2]
        target_vector = vectors[target][country][0]

        length_of_projection = (
            np.dot(target_vector, values["embedding"]) / 
            np.linalg.norm(values["embedding"])
        )

        distributions[target].append(length_of_projection)

        participants[participant]["projections"][target] = length_of_projection
        participants[participant]['gold'][target] = non_normalized_df.loc[participant][target]

with open('participants.pkl', 'wb') as f:
    pickle.dump(participants, f, protocol=pickle.HIGHEST_PROTOCOL)

participants['AT_50014']
