==> Codebase/9_evaluation.ipynb <==
import pandas as pd
import pickle
from sklearn.metrics import mean_absolute_error, cohen_kappa_score
from scipy.stats import spearmanr

with open('all_predictions.pkl', 'rb') as f:
    all_predictions = pickle.load(f)

results_df = pd.DataFrame(columns=['mean_absolute_error', 'cohen_kappa_score', 'spearmanr'])

for t, (y_true, y_pred) in all_predictions.items():
    results_df.loc[t] = [
        mean_absolute_error(y_true, y_pred),
        cohen_kappa_score(y_true, y_pred),
        spearmanr(y_true, y_pred)[0]  
    ]

results_df = results_df.T

results_df

results_df

results_df
