==> Codebase/10_unified_experiments.ipynb <==
import json
import pandas as pd
import pickle

from unified_experiments import run_experiment_for_target_variable

with open("codebook.json", "r") as codebook_file:
    codebook = json.load(codebook_file)

results_dict = {
    variable: (
        run_experiment_for_target_variable([variable]),
        codebook[variable]["scale_type"]
    )
    for variable in codebook.keys()
}

results_dict_dropped_missing_values = {
    variable: (
        run_experiment_for_target_variable([variable], drop_missing_val=True),
        codebook[variable]["scale_type"]
    )
    for variable in codebook.keys()
}

with open('results_dict.pkl', 'wb') as f:
    pickle.dump(results_dict, f, protocol=pickle.HIGHEST_PROTOCOL)

with open('results_dict_dropped_missing_values.pkl', 'wb') as f:
    pickle.dump(results_dict_dropped_missing_values, f, protocol=pickle.HIGHEST_PROTOCOL)

with open('results_dict.pkl', 'rb') as f:
    results_dict = pickle.load(f)

with open('results_dict_dropped_missing_values.pkl', 'rb') as f:
    results_dict_dropped_missing_values = pickle.load(f)

def big_dataframe(results_dict):
    rows = []

    for var_name, (df, var_type) in results_dict.items():
        variable_type = 'Unipolar' if var_type == 'u' else 'Bipolar'

        row = [variable_type, var_name] + df.iloc[0].tolist()
        rows.append(row)

    columns = ['Type', 'Variable'] + [col.replace('_', ' ').title() for col in df.columns.tolist()]

    results_df = pd.DataFrame(rows, columns=columns)
    results_df.set_index(['Type', 'Variable'], inplace=True)

    numeric_columns = results_df.select_dtypes(include='number').columns
    mean_df = results_df.groupby(['Type'])[numeric_columns].mean()

    return results_df, mean_df

def write_latex(df, path):
    with open(path, "w", encoding="utf-8") as f:
        f.write("\\begin{table}[H]\n")
        f.write(df.to_latex(bold_rows=True))
        f.write("\end{table}\n")

path = "../Reports/Final Report/tables"

results_df, mean_df = big_dataframe(results_dict)
write_latex(results_df, f"{path}/results_with_missing_values.tex")
write_latex(mean_df, f"{path}/means_with_missing_values.tex")

results_df2, mean_df2 = big_dataframe(results_dict_dropped_missing_values)
write_latex(results_df2, f"{path}/results_without_missing_values.tex")
write_latex(mean_df2, f"{path}/means_without_missing_values.tex")

results_df.describe()

results_df2.describe()
